+++ interface/resources/qml/hifi/commerce/common/sendAsset/images/send-money-effect-sm.jpg
