+++ interface/resources/qml/hifi/commerce/wallet/Wallet.qml
@@ -19,7 +19,7 @@ import "../../../styles-uit"
 import "../../../controls-uit" as HifiControlsUit
 import "../../../controls" as HifiControls
 import "../common" as HifiCommerceCommon
-import "./sendMoney"
+import "../common/sendAsset"
 
 Rectangle {
     HifiConstants { id: hifi; }
@@ -342,7 +342,7 @@ Rectangle {
         }
     }
 
-    SendMoney {
+    SendAsset {
         id: sendMoney;
         z: 997;
         visible: root.activeView === "sendMoney";
@@ -351,7 +351,7 @@ Rectangle {
         parentAppNavBarHeight: tabButtonsContainer.height;
 
         Connections {
-            onSendSignalToWallet: {
+            onSendSignalToParent: {
                 sendToScript(msg);
             }
         }
@@ -406,7 +406,7 @@ Rectangle {
     //
     Item {
         id: tabButtonsContainer;
-        visible: !needsLogIn.visible && root.activeView !== "passphraseChange" && root.activeView !== "securityImageChange" && sendMoney.currentActiveView !== "sendMoneyStep";
+        visible: !needsLogIn.visible && root.activeView !== "passphraseChange" && root.activeView !== "securityImageChange" && sendMoney.currentActiveView !== "sendAssetStep";
         property int numTabs: 5;
         // Size
         width: root.width;
