+++ scripts/system/commerce/wallet.js
@@ -11,11 +11,11 @@
 // See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
 
-/*global XXX */
+/* global getConnectionData */
 
 (function () { // BEGIN LOCAL_SCOPE
     Script.include("/~/system/libraries/accountUtils.js");
-    var request = Script.require('request').request;
+    Script.include("/~/system/libraries/connectionUtils.js");
 
     var MARKETPLACE_URL = Account.metaverseServerURL + "/marketplace";
 
@@ -51,94 +51,6 @@
     function sendToQml(message) {
         tablet.sendToQml(message);
     }
-
-    //***********************************************
-    //
-    // BEGIN Connection logic
-    //
-    //***********************************************
-    // Function Names:
-    //   - requestJSON
-    //   - getAvailableConnections
-    //   - getInfoAboutUser
-    //   - getConnectionData
-    //
-    // Description:
-    //   - Update all the usernames that I am entitled to see, using my login but not dependent on canKick.
-    var METAVERSE_BASE = Account.metaverseServerURL;
-    function requestJSON(url, callback) { // callback(data) if successfull. Logs otherwise.
-        request({
-            uri: url
-        }, function (error, response) {
-            if (error || (response.status !== 'success')) {
-                print("Error: unable to get", url, error || response.status);
-                return;
-            }
-            callback(response.data);
-        });
-    }
-    function getAvailableConnections(domain, callback) { // callback([{usename, location}...]) if successful. (Logs otherwise)
-        url = METAVERSE_BASE + '/api/v1/users?per_page=400&'
-        if (domain) {
-            url += 'status=' + domain.slice(1, -1); // without curly braces
-        } else {
-            url += 'filter=connections'; // regardless of whether online
-        }
-        requestJSON(url, function (connectionsData) {
-            callback(connectionsData.users);
-        });
-    }
-    function getInfoAboutUser(specificUsername, callback) {
-        url = METAVERSE_BASE + '/api/v1/users?filter=connections'
-        requestJSON(url, function (connectionsData) {
-            for (user in connectionsData.users) {
-                if (connectionsData.users[user].username === specificUsername) {
-                    callback(connectionsData.users[user]);
-                    return;
-                }
-            }
-            callback(false);
-        });
-    }
-    function getConnectionData(specificUsername, domain) { 
-        function frob(user) { // get into the right format
-            var formattedSessionId = user.location.node_id || '';
-            if (formattedSessionId !== '' && formattedSessionId.indexOf("{") != 0) {
-                formattedSessionId = "{" + formattedSessionId + "}";
-            }
-            return {
-                sessionId: formattedSessionId,
-                userName: user.username,
-                connection: user.connection,
-                profileUrl: user.images.thumbnail,
-                placeName: (user.location.root || user.location.domain || {}).name || ''
-            };
-        }
-        if (specificUsername) {
-            getInfoAboutUser(specificUsername, function (user) {
-                if (user) {
-                    updateUser(frob(user));
-                } else {
-                    print('Error: Unable to find information about ' + specificUsername + ' in connectionsData!');
-                }
-            });
-        } else {
-            getAvailableConnections(domain, function (users) {
-                if (domain) {
-                    users.forEach(function (user) {
-                        updateUser(frob(user));
-                    });
-                } else {
-                    sendToQml({ method: 'updateConnections', connections: users.map(frob) });
-                }
-            });
-        }
-    }
-    //***********************************************
-    //
-    // END Connection logic
-    //
-    //***********************************************
 
     //***********************************************
     //
@@ -678,7 +590,7 @@
                 }
                 removeOverlays();
                 break;
-            case 'sendMoney_sendPublicly':
+            case 'sendAsset_sendPublicly':
                 deleteSendMoneyParticleEffect();
                 sendMoneyRecipient = message.recipient;
                 var amount = message.amount;
